+++ app/build.gradle
@@ -36,4 +36,5 @@ dependencies {
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
     implementation 'com.github.bumptech.glide:glide:4.9.0'
     implementation 'com.google.android.gms:play-services-maps:16.1.0'
+    implementation 'com.android.support:recyclerview-v7:28.0.0'
 }
