--- app/src/main/res/layout/activity_catalogo.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout android:layout_height="match_parent"
-    android:layout_width="match_parent"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <ImageView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/iv_servicio_catalogo"
-        android:layout_centerHorizontal="true"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Salones que te brindan este servicio en Cali"
-        android:layout_below="@id/iv_servicio_catalogo"
-        android:id="@+id/tv_salones_catalogo"/>
-
-    <android.support.v7.widget.RecyclerView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@id/tv_salones_catalogo"
-        android:id="@+id/lista_salones_catalogo">
-
-    </android.support.v7.widget.RecyclerView>
-
-</RelativeLayout>
