+++ app/src/main/res/layout/fragment_citas.xml
@@ -5,10 +5,6 @@
     android:layout_height="match_parent"
     tools:context=".CitasFragment">
 
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="HELLO CITAS" />
+
 
 </FrameLayout>
