--- app/src/main/res/layout/renglon_salon.xml
@@ -1,99 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout android:layout_height="match_parent"
-    android:layout_width="match_parent"
-    android:layout_marginLeft="16dp"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-
-    <ImageView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/salon_test18"
-        android:id="@+id/imgsalon_renglon_salon"
-        android:layout_marginRight="16dp"
-        />
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:layout_toRightOf="@+id/imgsalon_renglon_salon"
-        android:layout_marginTop="10dp"
-        android:id="@+id/ll_nomycalif_salon">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="test18"
-        android:layout_gravity="center_horizontal"
-        android:textSize="20sp"
-        android:id="@+id/tv_salon_renglon_salon"/>
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_toRightOf="@id/imgsalon_renglon_salon"
-        >
-
-        <ImageView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/ic_calificacion_llena_25x25"/>
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_calificacion_llena_25x25"/>
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_calificacion_llena_25x25"/>
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_calificacion_llena_25x25"/>
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_calificacion_25x25"/>
-
-
-    </LinearLayout>
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:layout_marginLeft="24dp"
-        android:layout_toRightOf="@id/ll_nomycalif_salon"
-        >
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="24dp"
-            android:text="¿Como llegar?"
-            android:textSize="12sp"
-            android:layout_marginBottom="8dp"
-            android:textColor="#ffff"
-            android:layout_gravity="center_horizontal"
-            android:background="@drawable/btn_circular_secundario"/>
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="32dp"
-            android:text="Agenda tu cita"
-            android:padding="4dp"
-            android:textSize="14sp"
-            android:textColor="#ffff"
-            android:background="@drawable/btn_circular_principal"/>
-
-
-    </LinearLayout>
-
-
-</RelativeLayout>
